hey-25: it should make sure that only the person who has created the question can archive the question

+++ app/Http/Controllers/QuestionController.php
@@ -53,6 +53,8 @@ class QuestionController extends Controller
 
     public function archive(Question $question): RedirectResponse
     {
+        $this->authorize('archive', $question);
+
         $question->delete();
 
         return back();
